pyproject.toml
[tool.poetry]
name = "authen_docker"
version = "0.1.0"
description = ""
authors = ["NestorAlban <albquenes7294@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
python-multipart = "^0.0.5"
SQLAlchemy = "^1.4.37"
psycopg2 = "^2.9.3"
python-dotenv = "^0.20.0"
passlib = "^1.7.4"
bcrypt = "^3.2.2"
python-jose = "^3.3.0"
fastapi-mail = "^1.0.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
